+++ src/app/app.component.ts
@@ -143,7 +143,9 @@ export class AppComponent implements OnInit {
   ) {
 
     this.saveInStorageNumberOfOpenedChatTab();
-    this.getPageState();
+
+    this.listenToUrlChanges();
+    // this.getPageState();
 
     // location.subscribe((val) => {
 
@@ -160,46 +162,48 @@ export class AppComponent implements OnInit {
     //   }
     // });
 
+
+
+
+  }
+
+  listenToUrlChanges() {
+    const self = this;
     window.addEventListener('hashchange', function () {
-      console.log('location changed!');
+      // console.log('location changed!');
+
       const convId = getParameterByName('convId')
-      console.log('[APP-COMP] ngOnInit convId get with getParameterByName  ', convId)
-   
+      // console.log('[APP-COMP] getParameterByName convId ', convId)
       if (convId) {
         setTimeout(() => {
-          events.publish('convid:haschanged', convId);
+          self.events.publish('supportconvid:haschanged', convId);
         }, 0);
       }
-    
 
-      // this.navigateToDetail(convId, requesterFullaname)
+      const contact_id = getParameterByName('contact_id')
+      // console.log('[APP-COMP] getParameterByName contact_id ', contact_id)
+      const contact_fullname = getParameterByName('contact_fullname')
+      // console.log('[APP-COMP] getParameterByName contact_fullname ', contact_fullname)
+      if (contact_id && contact_fullname) {
+        setTimeout(() => {
+          self.router.navigateByUrl('conversation-detail/' + contact_id + '/' + contact_fullname + '/new');
+          self.events.publish('directconvid:haschanged', contact_id);
+        }, 0);
+        
+      } else {
+        // console.log('[APP-COMP] contact_id and contact_fullname are null')
+      }
 
-    
-
-    })
-
-
-    // router.events.subscribe((val) => {
-    //   this.checkCurrentURL();
-    //  });
-    // FOR TEST
-    // const last_project = { "user_available": true, "number_assigned_requests": 59, "last_login_at": "2021-08-09T17:30:55.234Z", "status": "active", "_id": "6112bc8f58c958003495a2cb", "id_project": { "status": 100, "_id": "60ffe291f725db00347661ef", "name": "27-LUGLIO-21-STRIPE-TEST", "activeOperatingHours": false, "createdBy": "608ad02d3a4dc000344ade17", "profile": { "name": "pro", "trialDays": 30, "agents": 5, "type": "payment", "subStart": "2021-11-18T10:42:41.000Z", "subEnd": "2021-11-19T10:42:41.000Z", "subscriptionId": "sub_Jvf4kABe9t8JvX", "last_stripe_event": "invoice.payment_succeeded" }, "versions": 20115, "channels": [{ "name": "chat21" }], "createdAt": "2021-07-27T10:40:17.752Z", "updatedAt": "2021-11-18T11:55:01.346Z", "__v": 0, "widget": { "preChatForm": true, "preChatFormJson": [{ "name": "userFullname", "type": "text", "mandatory": true, "label": { "en": "Your name", "it": "Il tuo nome" } }, { "name": "userEmail", "type": "text", "mandatory": true, "regex": "/^(?=.{1,254}$)(?=.{1,64}@)[-!#$%&'*+/0-9=?A-Z^_`a-z{|}~]+(.[-!#$%&'*+/0-9=?A-Z^_`a-z{|}~]+)*@[A-Za-z0-9]([A-Za-z0-9-]{0,61}[A-Za-z0-9])?(.[A-Za-z0-9]([A-Za-z0-9-]{0,61}[A-Za-z0-9])?)+$/", "label": { "en": "Your email", "it": "La tua email" }, "errorLabel": { "en": "Invalid email address", "it": "Indirizzo email non valido" } }, { "name": "tel", "mandatory": true, "label": { "en": "Your phone number", "it": "Il tuo numero di telefono" } }], "preChatFormCustomFieldsEnabled": true }, "trialExpired": true, "trialDaysLeft": 84, "isActiveSubscription": true, "id": "60ffe291f725db00347661ef" }, "id_user": "60aa0fef1482fe00346854a7", "role": "admin", "createdBy": "608ad02d3a4dc000344ade17", "createdAt": "2021-08-10T17:51:11.318Z", "updatedAt": "2021-11-19T08:08:21.437Z", "__v": 0, "presence": { "status": "online", "changedAt": "2021-11-19T08:08:21.432Z" }, "isAuthenticated": true, "id": "6112bc8f58c958003495a2cb" }
-    // localStorage.setItem('last_project', JSON.stringify(last_project))
+      const conversation_detail = getParameterByName('conversation_detail')
+      // console.log('[APP-COMP] getParameterByName conversation_detail ', conversation_detail)
+      if (conversation_detail) {
+        setTimeout(() => {
+          self.router.navigate(['conversation-detail/'])
+        }, 0);
+      } 
+    });
   }
-  navigateToDetail(convId, requesterFullaname) {
-    console.log('[APP-COMP] navigateToDetail  ');
-    let pageUrl = 'conversation-detail/' + convId + '/' + requesterFullaname + '/active';
-    this.router.navigateByUrl(pageUrl);
-    this.location.replaceState('conversation-detail/' + convId + '/' + requesterFullaname + '/active');
-  }
 
-
-  // private checkCurrentURL() {
-  //   console.log("location : " + window.location.pathname)
-  //   const convId = getParameterByName('convId')
-  //   console.log('[APP-COMP] ngOnInit convId get with getParameterByName  ', convId)
-  //   this.events.publish('convid:haschanged', convId);
-  // }
   saveInStorageNumberOfOpenedChatTab() {
     this.logger.log('Calling saveInStorageChatOpenedTab!');
 
@@ -213,12 +217,13 @@ export class AppComponent implements OnInit {
     }
     const terminationEvent = 'onpagehide' in self ? 'pagehide' : 'unload';
     window.addEventListener(terminationEvent, (event) => {
-      localStorage.tabCount = +localStorage.tabCount - 1;
+      if (localStorage.tabCount > 0) {
+        localStorage.tabCount = +localStorage.tabCount - 1;
+      }
     }, { capture: true });
   }
 
   getPageState() {
-
     const getState = () => {
       localStorage.setItem('visibilityState', document.visibilityState)
       if (document.visibilityState === 'hidden') {
@@ -316,17 +321,7 @@ export class AppComponent implements OnInit {
     this.listenToPostMsgs();
   }
 
-  // ngAfterViewInit() {
-  //   console.log('[APP-COMP] ngAfterViewInit')
-  // }
-
-  // ngDoCheck(){
-  //   console.log('[APP-COMP] ngDoCheck')
-  // }
-
-
-
-
+ 
   listenToPostMsgs() {
     window.addEventListener("message", (event) => {
       // console.log("[APP-COMP] message event ", event);
@@ -736,7 +731,7 @@ export class AppComponent implements OnInit {
   }
 
   checkPlatform() {
-    this.logger.debug('[APP-COMP] checkPlatform');
+   console.log('[APP-COMP] checkPlatform');
     // let pageUrl = '';
     // try {
     //   const pathPage = this.route.snapshot.firstChild.routeConfig.path;
@@ -752,16 +747,21 @@ export class AppComponent implements OnInit {
     // }
 
     if (checkPlatformIsMobile()) {
+      // console.log('[APP-COMP] checkPlatformIsMobile',checkPlatformIsMobile());
       this.platformIs = PLATFORM_MOBILE;
       const IDConv = this.route.snapshot.firstChild.paramMap.get('IDConv');
-      this.logger.log('[APP-COMP] PLATFORM_MOBILE2 navigateByUrl', PLATFORM_MOBILE, this.route.snapshot);
+      // console.log('[APP-COMP]  platformIs', this.platformIs);
+      // console.log('[APP-COMP] PLATFORM', PLATFORM_MOBILE, 'route.snapshot', this.route.snapshot);
       if (!IDConv) {
         this.router.navigateByUrl('conversations-list')
       }
       // this.router.navigateByUrl(pageUrl);
       // this.navService.setRoot(ConversationListPage, {});
     } else {
+      // console.log('[APP-COMP] checkPlatformIsMobile',checkPlatformIsMobile());
       this.platformIs = PLATFORM_DESKTOP;
+      // console.log('[APP-COMP]  platformIs', this.platformIs);
+      // console.log('[APP-COMP] PLATFORM', PLATFORM_DESKTOP, 'route.snapshot',  this.route.snapshot);
       this.logger.log('[APP-COMP] PLATFORM_DESKTOP ', this.navService);
 
       this.navService.setRoot(ConversationListPage, {});
@@ -911,7 +911,7 @@ export class AppComponent implements OnInit {
 
     this.conversationsHandlerService.conversationChanged.subscribe((conversation: ConversationModel) => {
 
-      console.log('[APP-COMP] ***** subscribeConversationChanged conversation: ', conversation);
+      // console.log('[APP-COMP] ***** subscribeConversationChanged conversation: ', conversation);
       const currentUser = JSON.parse(this.appStorageService.getItem('currentUser'));
       this.logger.log('[APP-COMP] ***** subscribeConversationChanged current_user: ', currentUser);