+++ src/app/services/app-config.ts
@@ -30,7 +30,7 @@ export class AppConfigProvider {
 
     return this.http.get(this.appConfig.remoteConfigUrl)
       .toPromise().then((data: any ) => {
-        console.log('AppConfigService loadAppConfig data: ', data);
+        // console.log('AppConfigService loadAppConfig data: ', data);
  
         const allconfig = data
         // console.log('[APP-CONFIG-SERVICE] - loadAppConfig allconfig: ', allconfig);
@@ -44,44 +44,46 @@ export class AppConfigProvider {
           if (wsUrlRelIsEmpty === false) {
             // console.log('[APP-CONFIG-SERVICE]- loadAppConfig allconfig !!!! exist - SERVER_BASE_URL', data.apiUrl);
 
-            if (allconfig.apiUrl.indexOf("http://") !== -1) {
+            // if (allconfig.apiUrl.indexOf("http://") !== -1) {
 
-              const ws_url = allconfig.apiUrl.replace("http://", "ws://").slice(0, -1) + allconfig.wsUrlRel;
-              // console.log('AppConfigService loadAppConfig allconfig !!!! exist - SERVER_BASE_URL protocol is HTTP - wsUrl', ws_url);
+            //   const ws_url = allconfig.apiUrl.replace("http://", "ws://").slice(0, -1) + allconfig.wsUrlRel;
+            //   // console.log('AppConfigService loadAppConfig allconfig !!!! exist - SERVER_BASE_URL protocol is HTTP - wsUrl', ws_url);
 
-              allconfig.wsUrl = ws_url
+            //   allconfig.wsUrl = ws_url
 
-            } else if (allconfig.apiUrl.indexOf("https://") !== -1) {
+            // } else if (allconfig.apiUrl.indexOf("https://") !== -1) {
 
-              const ws_url = allconfig.apiUrl.replace("https://", "wss://").slice(0, -1) + allconfig.wsUrlRel;
+            //   const ws_url = allconfig.apiUrl.replace("https://", "wss://").slice(0, -1) + allconfig.wsUrlRel;
 
-              allconfig.wsUrl = ws_url
+            //   allconfig.wsUrl = ws_url
 
-              // console.log('AppConfigService loadAppConfig allconfig !!!! exist - SERVER_BASE_URL protocol is HTTPS - wsUrl', ws_url);
-            } else {
-              console.log('AppConfigService loadAppConfig allconfig !!!! exist - SERVER_BASE_URL !!! IS RELATIVE - window.location ', window.location);
+            //   // console.log('AppConfigService loadAppConfig allconfig !!!! exist - SERVER_BASE_URL protocol is HTTPS - wsUrl', ws_url);
+            // } else {
+            
+            
+              // console.log('AppConfigService loadAppConfig allconfig !!!! exist - SERVER_BASE_URL !!! IS RELATIVE - window.location ', window.location);
 
               // console.log(window.location)
 
               if (window.location.protocol === 'http:') {
-                allconfig.wsUrl = 'ws://' + window.location.hostname + '/ws/'
+                allconfig.wsUrl = 'ws://' + window.location.hostname + allconfig.wsUrlRel
 
               } else if (window.location.protocol === 'https:') {
 
-                allconfig.wsUrl = 'wss://' + window.location.hostname + '/ws/'
+                allconfig.wsUrl = 'wss://' + window.location.hostname + allconfig.wsUrlRel
               } else {
 
-                allconfig.wsUrl = 'ws://' + window.location.hostname + '/ws/'
+                allconfig.wsUrl = 'ws://' + window.location.hostname + allconfig.wsUrlRel
               }
-            }
+            // }
 
           } else {
-            console.log('[APP-CONFIG-SERVICE] loadAppConfig allconfig !!!! exist wsUrlRel but IS EMPTY');
+            // console.log('[APP-CONFIG-SERVICE] loadAppConfig allconfig !!!! exist wsUrlRel but IS EMPTY');
           }
 
         } else {
 
-          console.log('[APP-CONFIG-SERVICE] loadAppConfig allconfig !!!! does not exist wsUrlRel');
+          // console.log('[APP-CONFIG-SERVICE] loadAppConfig allconfig !!!! does not exist wsUrlRel');
         }
 
         this.appConfig = allconfig;